+++ src/MarkdownEditorPackage.cs
@@ -9,12 +9,10 @@ namespace MarkdownEditor
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
     [ProvideAutoLoad(UIContextGuids80.SolutionExists)]
     [ProvideOptionPage(typeof(Options), "Text Editor\\Markdown", "General", 101, 111, true, new[] { "markdown", "md" }, ProvidesLocalizedCategoryName = false)]
-    [Guid(_packageGuidString)]
+    [Guid(PackageGuids.guidPackageString)]
+    [ProvideMenuResource("Menus.ctmenu", 1)]
     public sealed class MarkdownEditorPackage : Package
     {
-        private const string _packageGuidString = "9ca64947-e9ca-4543-bfb8-6cce9be19fd6";
-        private static Guid _guid = new Guid(_packageGuidString);
-
         public static Options Options { get; private set; }
 
         protected override void Initialize()
@@ -22,6 +20,8 @@ namespace MarkdownEditor
             Options = (Options)GetDialogPage(typeof(Options));
 
             Logger.Initialize(this, Vsix.Name);
+            CopyAsHtmlCommand.Initialize(this);
+
             base.Initialize();
         }
     }